+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.0] - 2023-08-01
+
 ### Fixed
 
 - Filter hosted zone by zone name when trying to find hosted zone ID.
@@ -284,7 +286,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2022-03-04
 
-[Unreleased]: https://github.com/giantswarm/irsa-operator/compare/v0.17.1...HEAD
+[Unreleased]: https://github.com/giantswarm/irsa-operator/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/giantswarm/irsa-operator/compare/v0.17.1...v0.18.0
 [0.17.1]: https://github.com/giantswarm/irsa-operator/compare/v0.17.0...v0.17.1
 [0.17.0]: https://github.com/giantswarm/irsa-operator/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/giantswarm/irsa-operator/compare/v0.15.0...v0.16.0